Extract matmul_kernel_configStr kernels into a separate file
